improve blog and contact

+++ modules/blog/blog.php
@@ -12,21 +12,41 @@ require_once 'ressources/php_markdown_lib_1.4.1/Michelf/Markdown.inc.php';?>
  	<h1 class="page-title">Blog</h1><hr/>
 
 <?php
-if (isset($_SESSION['valid_user'])) {
-	echo 'bravo';}
-else { exit;}
 include(DB);
-$query = 'SELECT e.title, e.entry, c.name, e.entry_id FROM entries e LEFT JOIN Category c ON e.category_id = c.id ORDER BY e.date_entered DESC';
+$query = 'SELECT e.title, e.entry, c.name, e.entry_id, e.isPublic, e.date_entered FROM entries e LEFT JOIN Category c ON e.category_id = c.id ORDER BY e.date_entered DESC';
  if ($r = mysql_query($query,$dbc)) {
 
 	while ($row = mysql_fetch_array($r)) {
 		// $markdown = new Markdown();
-		$entry = \Michelf\Markdown::defaultTransform($row['entry']);
-		print "<div class='panel'><h3><a href=\"/posts/{$row['entry_id']}\">{$row['title']}</a><span style='float:right;' class='label radius'>";
-		if (isset($row['name'])) echo $row['name'];
-		print "</span></h3> {$entry}<br />
-		<a href=\"index.php?p=edit_entry&id={$row['entry_id']}\">Edit</a>
-		<a href=\"delete_entry.php? id={$row['entry_id']}\">Delete</a></div>\n";
+		if (isset($_SESSION['valid_user'])) {
+			$entry = \Michelf\Markdown::defaultTransform($row['entry']);
+			$format = 'Y-m-d H:i:s';
+			$date_entry = date_create_from_format($format, $row['date_entered']);
+			// echo $date_entry->format('d/m/Y');
+			print "<div class='panel multiple-post'>
+					<h3>
+						<a href=\"/posts/{$row['entry_id']}\">{$row['title']}</a><span style='float:right;' class='label radius'>";
+			if (isset($row['name'])) echo $row['name'];
+			if ($row['isPublic'] == 0) echo ' PRIVE';
+			print "</span></h3>".$date_entry->format('d/m/Y').$entry;
+			if (isset($_SESSION['valid_user'])) {
+				print "<a href=\"/index.php?p=edit_entry&id={$row['entry_id']}\">Edit</a>
+					  <a href=\"/index.php?p=delete_entry&id={$row['entry_id']}\">Delete</a></div>\n";
+			}
+		}
+		else { 
+			if ($row['isPublic'] == 1){
+				$entry = \Michelf\Markdown::defaultTransform($row['entry']);
+				print "<div class='panel'>
+						<h3>
+							<a href=\"/posts/{$row['entry_id']}\">{$row['title']}</a><span style='float:right;' class='label radius'>";
+				if (isset($row['name'])) echo $row['name'];
+				if ($row['isPublic'] == 0) echo ' PRIVE';
+				print "</span></h3> {$entry}<br />
+				<a href=\"index.php?p=edit_entry&id={$row['entry_id']}\">Edit</a>
+				<a href=\"delete_entry.php? id={$row['entry_id']}\">Delete</a></div>\n";
+			}
+		}
 	}
 } 
 else { // Query didn't run.
